Add files via upload
This Work is a Capstone project of the Google Data Analytics certificate program hosted on Coursera. In this project we havedata of fitbit smart device users over a 30 days period. I analyzed the data collated and made recomendations for Bellabeats( a startup wellness company) on how to engage more with users and better market it's products.